Restructured the project file layout
