+++ PokerHands/Services/PlayerHand.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using PokerHands.Models;
 
-namespace PokerHands.Models
+namespace PokerHands.Services
 {
     public class PlayerHand
     {
