+++ PokerHands/Tests/Services/PlayerHandTests.cs
@@ -1,8 +1,9 @@
 ﻿using FluentAssertions;
 using PokerHands.Models;
+using PokerHands.Services;
 using Xunit;
 
-namespace PokerHands.Tests.Models
+namespace PokerHands.Tests.Services
 {
     public class PlayerHandTests
     {
